1. rename questionnaire setups appropriately. 2. create common enum model for stuff like gender, status etc.
