+++ src/app/setups/categories/clinics/clinics.component.html
@@ -77,7 +77,7 @@
 
   <nz-tab nzTitle="Questionnaire">
     <ng-template nz-tab>
-      <app-consultation-questions></app-consultation-questions>
+      <app-consultation-questionnaire-setup></app-consultation-questionnaire-setup>
     </ng-template>
   </nz-tab>
 </nz-tabset>
