+++ src/app/setups/setups.module.ts
@@ -69,7 +69,7 @@ import { MedicineRoutesComponent } from './pages/medicine-routes/medicine-routes
 import { ProductFormAndUnitComponent } from './pages/product-form-and-unit/product-form-and-unit.component';
 import { ProductTypeAndCategoryComponent } from './pages/product-type-and-category/product-type-and-category.component';
 import { StoreSetupComponent } from './categories/store-setup/store-setup.component';
-import {ConsultationQuestionsComponent} from "./pages/consultation-questions/consultation-questions.component";
+import {ConsultationQuestionnaireSetupComponent} from "./pages/consultation-questionnaire-setup/consultation-questionnaire-setup.component";
 
 @NgModule({
   declarations: [
@@ -137,7 +137,7 @@ import {ConsultationQuestionsComponent} from "./pages/consultation-questions/con
     LabSampleTypesComponent,
     ParameterServiceMapComponent,
     SampleTypeServiceMapComponent,
-    ConsultationQuestionsComponent
+    ConsultationQuestionnaireSetupComponent
   ],
   imports: [
     CommonModule,
