+++ src/app/shared/models/common.model.ts
@@ -1,0 +1,12 @@
+
+export enum Status {
+  ACTIVE = 'ACTIVE', INACTIVE = 'INACTIVE'
+}
+
+export enum Gender {
+  MALE = 'MALE', FEMALE = 'FEMALE', ALL = 'MALE,FEMALE'
+}
+
+export enum ValueType {
+  TEXT = 'Text', NUMBER = 'Number', BOOLEAN = 'True/False', SELECT = "Select"
+}
